refactor(desktop): 重构Live2D渲染适配双后端并优化文件服务
1.  /desktop/src/petcanvas.h:39 替换float为qreal作为旋转角度类型
2.  /desktop/src/main.cpp:1-14 添加USE_QT_WEBVIEW宏条件编译WebView初始化代码
3.  /desktop/BackPet.pro:10-41 重构WebEngine和WebView依赖检查逻辑
4.  /desktop/src/components/live2dwidget.h:2-81 重构为支持QWebEngineView和QQuickWidget双渲染后端
5.  /desktop/src/components/localfileserver.cpp:55-145 重构文件服务逻辑，支持qrc资源代理和修复路由警告
6.  /desktop/src/components/live2dwidget.cpp:4-424 实现双后端渲染逻辑和JS控制台转发
7.  /desktop/resources/live2d-view.html:103-710 优化Live2D参数处理逻辑，兼容更多模型参数范围

+++ desktop/src/petcanvas.h
@@ -39,7 +39,7 @@ struct Sphere3D {
     QVector3D pos;      // 3D 位置
     QVector3D scale;    // 缩放 (非等比 -> 椭球)
     QColor color;       // 颜色
-    float rotation;     // 旋转角度
+    qreal rotation;     // 旋转角度
     QVector3D rotAxis;  // 旋转轴
 };
